Add botanist functions to load file.
