+++ pipeline/load.py
@@ -100,6 +100,75 @@ def get_plant_id(connection, plant_data: dict) -> dict:
     return result
 
 
+def get_botanist_id(connection, plant_data: dict) -> dict:
+    """Gets the corresponding botanist ID from database using botanist name."""
+
+    curs = get_db_cursor(connection)
+
+    curs.execute("""SELECT botanist_id
+                    FROM botanist
+                    WHERE botanist_name = ?
+                    AND email = ?""",
+                 (plant_data["botanist"]["name"],
+                  plant_data["botanist"]["email"]))
+    result = curs.fetchone()[0]
+
+    return result
+
+
+def load_botanist_assignment_data(connection, plants_data: list[dict]) -> None:
+    """Loads botanist assignment data from dictionary to botanist assignment table in database."""
+
+    insert_query = """
+                IF NOT EXISTS (
+                    SELECT 1 FROM botanist_assignment
+                    WHERE botanist_id = ?
+                    AND plant_id = ?)
+                BEGIN
+                    INSERT INTO botanist_assignment
+                    VALUES (?, ?)
+                END
+                """
+
+    curs = get_db_cursor(connection)
+    for plant in plants_data:
+        plant_id = get_plant_id(connection, plant)
+        botanist_id = get_botanist_id(connection, plant)
+        curs.execute(
+            insert_query, (botanist_id,
+                           plant_id,
+                           botanist_id,
+                           plant_id)
+        )
+        connection.commit()
+
+
+def load_botanist_data(connection, plants_data: list[dict]) -> None:
+    """Loads botanist data from dictionary to botanist table in database."""
+
+    insert_query = """
+                IF NOT EXISTS (
+                    SELECT 1 FROM botanist
+                    WHERE botanist_name = ?
+                    AND email = ?)
+                BEGIN
+                    INSERT INTO botanist
+                    VALUES (?, ?, ?)
+                END
+                """
+
+    curs = get_db_cursor(connection)
+    for plant in plants_data:
+        botanist = plant["botanist"]
+        curs.execute(
+            insert_query, (botanist["name"],
+                           botanist["email"],
+                           botanist["name"],
+                           botanist["email"],
+                           botanist["phone"]))
+        connection.commit()
+
+
 def load_plant_master_data(connection, plants_data: list[dict]) -> None:
     """Loads plant master data from dictionary to plant table in SQL Server database."""
 
@@ -214,6 +283,8 @@ if __name__ == "__main__":
 
     conn = get_db_connection()
 
+    load_botanist_data(conn, seed_data)
+    load_botanist_assignment_data(conn, seed_data)
     load_country_data(conn, seed_data)
     load_origin_data(conn, seed_data)
     load_plant_master_data(conn, seed_data)
